Remove some of the flash messages
I deleted some of the flash messages because I did not need it.
Also when turbo is active, flash messages don't show up in the page.
I could have changed the respond format and make it work, but
this requires turbo streams, and right now I don't use streams but frames.
I might change this approach in the future when I start using turbo streams.