Add back deleted posts_controller method
I added back the delete code which was responsible for returning all
posts of all the current_user’s friends along with the curren_user’s post.
Because, previous method had a edge case in real life practice.
I decided that trading stability with the performance does not worth it.
Otherwise, user mightn't see any posts in Feed if nonfriends share posts too much.
This could've happened if more than 100 of posts are shared and non of the
posts belong to the current_user or the current_users' friends.
This is why I added the 'good old' code back :)